requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile requirements.in
#
blinker==1.9.0
    # via flask
certifi==2025.4.26
    # via requests
charset-normalizer==3.4.2
    # via requests
click==8.2.1
    # via flask
colorama==0.4.6
    # via click
dash==3.0.4
    # via -r requirements.in
flask==3.0.3
    # via dash
idna==3.10
    # via requests
importlib-metadata==8.7.0
    # via dash
itsdangerous==2.2.0
    # via flask
jinja2==3.1.6
    # via flask
markupsafe==3.0.2
    # via
    #   jinja2
    #   werkzeug
narwhals==1.42.1
    # via plotly
nest-asyncio==1.6.0
    # via dash
numpy==2.3.0
    # via
    #   pandas
    #   pandas-stubs
packaging==25.0
    # via plotly
pandas==2.3.0
    # via -r requirements.in
pandas-stubs==2.2.3.250527
    # via -r requirements.in
plotly==6.1.2
    # via
    #   -r requirements.in
    #   dash
python-dateutil==2.9.0.post0
    # via pandas
python-dotenv==1.1.0
    # via -r requirements.in
pytz==2025.2
    # via pandas
requests==2.32.4
    # via
    #   -r requirements.in
    #   dash
retrying==1.3.4
    # via dash
six==1.17.0
    # via
    #   python-dateutil
    #   retrying
types-pytz==2025.2.0.20250516
    # via
    #   -r requirements.in
    #   pandas-stubs
typing-extensions==4.14.0
    # via dash
tzdata==2025.2
    # via pandas
urllib3==2.4.0
    # via requests
werkzeug==3.0.6
    # via
    #   dash
    #   flask
zipp==3.23.0
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools
